+++ modules/code-editor/src/views/full/FileNavigator.tsx
@@ -1,4 +1,15 @@
-import { Classes, Icon, ITreeNode, Position, Tooltip, Tree } from '@blueprintjs/core'
+import {
+  Classes,
+  ContextMenu,
+  Icon,
+  ITreeNode,
+  Menu,
+  MenuDivider,
+  MenuItem,
+  Position,
+  Tooltip,
+  Tree
+} from '@blueprintjs/core'
 import { observe } from 'mobx'
 import { inject, observer } from 'mobx-react'
 import React from 'react'
@@ -6,7 +17,6 @@ import ReactDOM from 'react-dom'
 
 import { EditableFile } from '../../backend/typings'
 
-import { showExampleContextMenu, showModuleConfigContextMenu, showStandardContextMenu } from './components/ContextMenus'
 import { RootStore, StoreDef } from './store'
 import { EditorStore } from './store/editor'
 import { buildTree, EXAMPLE_FOLDER_LABEL, FOLDER_EXAMPLE, FOLDER_ICON } from './utils/tree'
@@ -130,14 +140,72 @@ class FileNavigator extends React.Component<Props, State> {
     const file = node.nodeData as EditableFile
 
     if (this.props.contextMenuType === 'moduleConfig') {
-      return showModuleConfigContextMenu(file, e)
+      if (!file.botId) {
+        ContextMenu.show(
+          <Menu>
+            <MenuItem
+              id="btn-duplicateCurrent"
+              icon="duplicate"
+              text="Duplicate to current bot"
+              onClick={() => this.props.duplicateFile(file, { forCurrentBot: true, keepSameName: true })}
+            />
+          </Menu>,
+          { left: e.clientX, top: e.clientY }
+        )
+      } else {
+        ContextMenu.show(
+          <Menu>
+            <MenuItem id="btn-delete" icon="delete" text="Delete" onClick={() => this.props.deleteFile(file)} />
+          </Menu>,
+          { left: e.clientX, top: e.clientY }
+        )
+      }
+
+      return
     }
 
     if (file.isExample) {
-      return showExampleContextMenu(file, e)
+      ContextMenu.show(
+        <Menu>
+          <MenuItem
+            id="btn-duplicateCurrent"
+            icon="duplicate"
+            text={file.type === 'action' ? 'Copy example to my bot' : 'Copy example to global hooks'}
+            onClick={() =>
+              this.props.duplicateFile(file, { forCurrentBot: file.type === 'action', keepSameName: true })
+            }
+          />
+        </Menu>,
+        { left: e.clientX, top: e.clientY }
+      )
+      return
     }
 
-    return showStandardContextMenu(file, e, node)
+    const isDisabled = file.name.startsWith('.')
+    ContextMenu.show(
+      <Menu>
+        <MenuItem id="btn-rename" icon="edit" text="Rename" onClick={() => this.renameTreeNode(node)} />
+        <MenuItem id="btn-delete" icon="delete" text="Delete" onClick={() => this.props.deleteFile(file)} />
+        <MenuDivider />
+        <MenuItem id="btn-duplicate" icon="duplicate" text="Duplicate" onClick={() => this.props.duplicateFile(file)} />
+        <MenuDivider />
+        <MenuItem
+          id="btn-enable"
+          icon="endorsed"
+          text="Enable"
+          disabled={!isDisabled}
+          onClick={() => this.props.enableFile(file)}
+        />
+        <MenuItem
+          id="btn-disable"
+          icon="disable"
+          text="Disable"
+          disabled={isDisabled}
+          onClick={() => this.props.disableFile(file)}
+        />
+      </Menu>,
+      { left: e.clientX, top: e.clientY }
+    )
   }
 
   renameTreeNode = async (node: ITreeNode) => {
